steel_core::behavior

Module block

Block behavior trait and registry.

Re-exports

pub use context::BlockCollisionBoxes;

pub use context::BlockCollisionContext;

pub use context::BlockEntityCreation;

pub use context::BlockLootContext;

pub use context::EntityFallDamage;

pub use context::EntityFallOnContext;

pub use context::EntityFallOnFacts;

pub use context::EntityLandingContext;

pub use context::Fallable;

pub use context::PickupResult;

pub use context::RailBehavior;

pub use registry::BlockBehaviorRegistry;

pub use registry::DefaultBlockBehavior;

Modules

collision 🔒

context 🔒

registry 🔒

waterlogging 🔒

Structs

BrushableData

Data exposed by blocks that support vanilla archaeology brushing.

Traits

BlockBehavior

Trait defining the behavior of a block.

Functions

default_can_be_replaced 🔒

Vanilla BlockBehaviour.canBeReplaced(BlockState, BlockPlaceContext).

drop_from_block_interact_loot_table 🔒

Gets random loot from a given loot table reference and other factors, and returns each item from it in a Vec.

try_drop_experience 🔒

Samples and applies enchantment effects to a block experience drop.