steel_core::behavior::block

Function drop_from_block_interact_loot_table

pub(crate) fn drop_from_block_interact_loot_table(
    key: LootTableRef,
    interacted_block_state: BlockStateId,
    _interacted_block_entity: Option<SharedBlockEntity>,
    tool: Option<&ItemStack>,
    interacting_entity: Option<&dyn Entity>,
    rng: &mut ThreadRng,
) -> Vec<ItemStack>

Gets random loot from a given loot table reference and other factors, and returns each item from it in a Vec.