steel_core::behavior::blocks::vegetation

Trait Vegetation

pub trait Vegetation {
    // Provided method
    fn may_place_on(
        &self,
        state: BlockStateId,
        _world: &dyn LevelReader,
        _pos: BlockPos,
    ) -> bool { ... }
}

Common behavior for vegetation blocks

Provided Methods

fn may_place_on( &self, state: BlockStateId, _world: &dyn LevelReader, _pos: BlockPos, ) -> bool

Checks if the vegetation block can be placed on the given block state below on the given position below.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".