steel_core::block_entity::entities

Struct SignBlockEntity

pub struct SignBlockEntity {
    level: Weak<World>,
    block_entity_type: BlockEntityTypeRef,
    pos: BlockPos,
    state: BlockStateId,
    removed: bool,
    pub front_text: SignText,
    pub back_text: SignText,
    pub is_waxed: bool,
    player_who_may_edit: Option<Uuid>,
}

Sign block entity.

Stores text on both front and back sides of the sign.

Fields
level: Weak<World>

Weak reference to the world for marking chunks dirty.

block_entity_type: BlockEntityTypeRef

Block entity type (sign or hanging_sign).

pos: BlockPos

Position in the world.

state: BlockStateId

Current block state.

removed: bool

Whether this entity has been marked for removal.

front_text: SignText

Text on the front side.

back_text: SignText

Text on the back side.

is_waxed: bool

Whether the sign is waxed (prevents editing).

player_who_may_edit: Option<Uuid>

UUID of the player currently allowed to edit this sign. Used to prevent multiple players from editing simultaneously.

Implementations

impl SignBlockEntity

pub fn new(level: Weak<World>, pos: BlockPos, state: BlockStateId) -> Self

Creates a new sign block entity.

pub fn new_hanging( level: Weak<World>, pos: BlockPos, state: BlockStateId, ) -> Self

Creates a new hanging sign block entity.

pub fn with_type( level: Weak<World>, block_entity_type: BlockEntityTypeRef, pos: BlockPos, state: BlockStateId, ) -> Self

Creates a sign block entity with a specific type.

pub const fn get_player_who_may_edit(&self) -> Option<Uuid>

Gets the UUID of the player currently allowed to edit this sign.

pub const fn set_player_who_may_edit(&mut self, player: Option<Uuid>)

Sets the player allowed to edit this sign.

pub fn is_other_player_editing(&self, player_uuid: Uuid) -> bool

Checks if another player (not the given one) is currently editing this sign.

pub const fn get_text(&self, front: bool) -> &SignText

Gets the text for a side.

pub const fn get_text_mut(&mut self, front: bool) -> &mut SignText

Gets mutable text for a side.

pub fn set_text(&mut self, text: SignText, front: bool)

Sets the text for a side.