steel_core::chunk

Module chunk_access

This module contains the ChunkAccess enum, which is used to access chunks in different states.

Enums

ChunkAccess

An enum that allows access to a chunk in different states.

ChunkStatus

The status of a chunk.