steel_core::chunk::chunk_access

Enum ChunkStatus

pub enum ChunkStatus {
    Empty,
    StructureStarts,
    StructureReferences,
    Biomes,
    Noise,
    Surface,
    Carvers,
    Features,
    InitializeLight,
    Light,
    Spawn,
    Full,
}

The status of a chunk.

Variants

Empty

The chunk is empty.

StructureStarts

The chunk is being processed for structure starts.

StructureReferences

The chunk is being processed for structure references.

Biomes

The chunk is being processed for biomes.

Noise

The chunk is being processed for noise.

Surface

The chunk is being processed for surfaces.

Carvers

The chunk is being processed for carvers.

Features

The chunk is being processed for features.

InitializeLight

The chunk is being initialized for light.

Light

The chunk is being lit.

Spawn

The chunk is being spawned.

Full

The chunk is fully generated.

Implementations

impl ChunkStatus

pub const fn get_index(self) -> usize

Gets the index of the status.

pub const fn from_index(index: usize) -> Option<Self>

Gets the status from an index.

impl ChunkStatus

pub const fn next(self) -> Option<Self>

Gets the next status in the generation order.

Panics

This function will panic if the chunk is at the Full status.

pub const fn parent(self) -> Option<Self>

Gets the parent status in the generation order.

pub const fn heightmaps_after(self) -> &'static [HeightmapType]

Returns the heightmap types that should be updated at this status.

Before CARVERS status, worldgen heightmaps are used. At CARVERS and after, final heightmaps are used.