steel_core::chunk

Module chunk_pyramid

This module contains the ChunkPyramid, which is used to check chunk dependencies.

Structs

Builder

A builder for a chunk step.

ChunkDependencies

A map of radius to the required chunk status.

ChunkPyramid

Represents the hierarchy and dependencies for chunk generation or loading.

ChunkPyramidBuilder

A builder for a chunk pyramid.

ChunkStep

A chunk step.

Statics

GENERATION_PYRAMID

The generation pyramid.

LOADING_PYRAMID

The loading pyramid.

Type Aliases

ChunkStatusTask

A task that generates a chunk.