steel_core::chunk::chunk_pyramid

Struct Builder

pub struct Builder {
    status: ChunkStatus,
    parent: Option<Arc<ChunkStep>>,
    direct_dependencies_by_radius: Vec<ChunkStatus>,
    block_state_write_radius: i32,
    task: ChunkStatusTask,
}

A builder for a chunk step.

Fields
status: ChunkStatusparent: Option<Arc<ChunkStep>>direct_dependencies_by_radius: Vec<ChunkStatus>block_state_write_radius: i32task: ChunkStatusTask
Implementations

impl Builder

pub fn new(status: ChunkStatus, parent: Option<Arc<ChunkStep>>) -> Self

Creates a new chunk step builder.

Panics

Panics if the status is not the next status of the parent.

pub fn add_requirement(self, status: ChunkStatus, radius: usize) -> Self

Adds a requirement to the step.

Panics

Panics if the status is greater than the current status.

pub fn block_state_write_radius(self, radius: i32) -> Self

Sets the block state write radius for the step.

pub fn set_task(self, task: ChunkStatusTask) -> Self

Sets the task for the step.

pub fn build(self) -> ChunkStep

Builds the chunk step.

fn build_accumulated_dependencies(&self) -> ChunkDependencies