steel_core::chunk

Module paletted_container

A paletted container is a container that can be either homogeneous or heterogeneous.

Structs

HeterogeneousPalette

A heterogeneous palette container.

Enums

PaletteMode 🔒

PalettedContainer

A paletted container.

Traits

ToGlobalId

A trait for converting a value to a global ID.

Functions

pack_bits 🔒

Type Aliases

BiomePalette

A palette container for biomes.

BlockPalette

A palette container for blocks.

Cube 🔒

3d array indexed by y,z,x