steel_core::chunk::paletted_container

Type Alias BlockPalette

pub type BlockPalette = PalettedContainer<BlockStateId, 16>;

A palette container for blocks.

Aliased Type

pub enum BlockPalette {
    Homogeneous(BlockStateId),
    Heterogeneous(HeterogeneousPalette<BlockStateId, 16>),
    Building(Box<[[[BlockStateId; 16]; 16]; 16]>),
}

Variants

Homogeneous(BlockStateId)

A homogeneous container, where all values are the same.

Heterogeneous(HeterogeneousPalette<BlockStateId, 16>)

A heterogeneous container, where values can be different.

Building(Box<[[[BlockStateId; 16]; 16]; 16]>)

Write-only build mode: raw cube without palette tracking. set is a single store; get is a direct read. Convert back via Self::finalize_building.

Implementations

impl BlockPalette

pub fn non_empty_block_count(&self) -> u16

Gets the number of non-empty blocks in the container.

pub fn has_only_air(&self) -> bool

Returns true if this palette contains only air blocks.