steel_core::chunk::world_gen_context

Struct WorldGenContext

pub struct WorldGenContext {
    pub generator: Arc<ChunkGeneratorType>,
    world: Weak<World>,
}

Context for world generation.

Similar to vanilla’s WorldGenContext, this provides access to the level/dimension and generation infrastructure.

Fields
generator: Arc<ChunkGeneratorType>

The chunk generator to use.

world: Weak<World>

Weak reference to the world (to avoid circular Arc reference). Use world() to get a strong reference when needed.

Implementations

impl WorldGenContext

pub fn new(generator: Arc<ChunkGeneratorType>, world: Weak<World>) -> Self

Creates a new WorldGenContext.

pub fn world(&self) -> Arc<World>

Gets a strong reference to the world.

Panics

Panics if the world has been dropped.

pub fn weak_world(&self) -> Weak<World>

Gets a weak reference to the world.

This is useful for passing to chunks without creating a strong reference cycle.

pub fn min_y(&self) -> i32

Returns the minimum Y coordinate of the world.

pub fn height(&self) -> i32

Returns the total height of the world in blocks.

pub fn section_count(&self) -> usize

How many sections this dimension has