steel_core::chunk::world_gen_context

Type Alias NetherGenerator

pub type NetherGenerator = VanillaGenerator<NetherNoises>;

Type alias for nether generator.

Aliased Type

pub struct NetherGenerator {
    biome_source: BiomeSourceKind,
    noises: Box<NetherNoises>,
    splitter: RandomSplitter,
    ore_veinifier: Option<OreVeinifier>,
    surface_system: SurfaceSystem,
    default_block_id: BlockStateId,
    biome_zoom_seed: i64,
    _phantom: PhantomData<NetherNoises>,
}

Fields
biome_source: BiomeSourceKind

Biome source for this dimension. Determines biomes at each quart position.

noises: Box<NetherNoises>

Noise generators for this dimension’s density functions. Boxed because noise structs can be large.

splitter: RandomSplitter

Seed positional splitter for per-chunk construction of aquifers.

ore_veinifier: Option<OreVeinifier>

Ore vein generator for replacing stone with ore blocks.

surface_system: SurfaceSystem

Surface system for biome-specific block replacement.

default_block_id: BlockStateId

Block state ID for the default block, cached at construction time.

biome_zoom_seed: i64

Obfuscated seed for BiomeManager biome zoom fuzzing.

_phantom: PhantomData<NetherNoises>