steel_core::command

Module arguments

This module contains types and utilities for parsing command arguments.

Modules

anchor

A anchor argument.

block_pos

A block position argument.

bool

A boolean argument.

domain

Argument that resolves a configured domain name.

enchantment

An enchantment argument

entity

A entity argument.

entity_type

Entity type command arguments.

float

A float argument.

gamemode

A gamemode argument.

integer

An integer argument.

item

An item argument

player

A player argument.

rotation

A rotation argument.

structure

A structure argument.

text_component

A text argument.

time

A time argument.

vector2

A vector2 argument.

vector3

A vector3 argument.

world

Argument that resolves a world identifier to a loaded world.

Structs

Helper 🔒

SuggestionContext

Context passed to suggestion methods containing previously parsed arguments.

Enums

ParsedValue

A parsed value that can be stored in suggestion context.

Traits

CommandArgument

A trait that defines a command argument parser.