steel_core::command::brigadier::suggestion

Trait SuggestionProvider

pub(crate) trait SuggestionProvider<S, A: CommandArgumentParser<S>>: Send + Sync {
    // Required method
    fn list_suggestions(
        &self,
        context: &ArgumentSuggestionContext<'_, S, A::Value>,
        builder: &mut SuggestionsBuilder<'_>,
    );
}

A provider to add suggestions to a builder.

Functions that have the same function signature as that of SuggestionProvider::list_suggestions also implement this trait.

Required Methods

fn list_suggestions( &self, context: &ArgumentSuggestionContext<'_, S, A::Value>, builder: &mut SuggestionsBuilder<'_>, )

Adds suggestions, according to this provider, to the given builder.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".