steel_core::entity::entities::objects

Module projectiles

Projectile entity implementations.

Modules

ender_pearl 🔒

Thrown ender pearl projectile entity (ThrownEnderpearl).

eye_of_ender 🔒

Eye of ender entity implementation (EyeOfEnderEntity).

firework_rocket 🔒

Firework rocket projectile entity (FireworkRocketEntity).

fishing_hook 🔒

snowball 🔒

Thrown snowball projectile entity (Snowball).

thrown_egg 🔒

Thrown egg projectile entity (ThrownEgg).

Structs

EnderPearlEntity

A thrown ender pearl.

EyeOfEnderEntity

A thrown eye of ender, seeking the nearest stronghold.

FireworkRocketEntity

A launched firework rocket.

FishingHookEntity

A fishing hook.

FishingHookState

This struct holds entity specific state information per fishing hook entity.

SnowballEntity

A thrown snowball.

ThrownEggEntity

A thrown egg.