steel_core::entity::entities::objects::projectiles

Module fishing_hook

Structs

FishingHookEntity

A fishing hook.

FishingHookState

This struct holds entity specific state information per fishing hook entity.

Enums

BobberState 🔒

Collection of possible states the fishing bobber of the FishingHookEntity can take on. Equivalent to Java’s FishingHook.FishHookState (we renamed for clarity)

OpenWaterType 🔒

Collection of possible types associated with open water.

Constants

DEGREE_180 🔒

DEGREE_360 🔒

DEG_TO_RAD 🔒

DMG_CAUGHT 🔒

DMG_DEFAULT 🔒

DMG_ITEM_ENTITY 🔒

DMG_ON_GROUND 🔒

FIVE_SECONDS 🔒

FOUR_SECONDS 🔒

MAX_DISTANCE_SQR 🔒

MAX_OUT_OF_WATER_TIME

ONE_MINUTE 🔒

ONE_SECOND 🔒

THIRTY_SECONDS 🔒

THREE_SECONDS 🔒

TWO_SECONDS 🔒