steel_core::entity::entities::objects

Module technical

Technical entity implementations.

Modules

interaction 🔒

marker 🔒

Structs

InteractionEntity

An invisible, invincible, interactable entity which records when a player clicks on its bounding box. It is used in map-making or data packs, and its bounding box is customizable.

InteractionEntityDataView

Provides an exclusive view of the synchronized entity data of an interaction entity. This includes its width, its height, and the response boolean.

MarkerEntity

An entity that is minimal. It cannot move, take damage, make sounds, and it doesn’t have any gameplay functions. Markers are not sent to the client, and are primarily intended for use in map-making and in data packs. This entity is purely server-side.

PlayerAction

Represents an action of a player. It contains the UUID of the player who executed this action and the timestamp (based on game time, expressed in ticks) when it was executed.