steel_core::entity::entities::objects::technical

Module interaction

Structs

InteractionEntity

An invisible, invincible, interactable entity which records when a player clicks on its bounding box. It is used in map-making or data packs, and its bounding box is customizable.

InteractionEntityDataView

Provides an exclusive view of the synchronized entity data of an interaction entity. This includes its width, its height, and the response boolean.

PlayerAction

Represents an action of a player. It contains the UUID of the player who executed this action and the timestamp (based on game time, expressed in ticks) when it was executed.

Constants

DEFAULT_DIMENSIONS 🔒

DEFAULT_HEIGHT 🔒

DEFAULT_RESPONSE 🔒

DEFAULT_WIDTH 🔒

TAG_ATTACK 🔒

TAG_HEIGHT 🔒

TAG_INTERACTION 🔒

TAG_PLAYER 🔒

TAG_RESPONSE 🔒

TAG_TIMESTAMP 🔒

TAG_WIDTH 🔒