steel_core::entity::projectile::throwable_item

Trait ThrowableItemProjectile

pub trait ThrowableItemProjectile: ThrowableProjectile {
    // Required methods
    fn get_default_item(&self) -> ItemRef;
    fn set_item(&self, item: ItemStack);
    fn get_item(&self) -> ItemStack;

    // Provided methods
    fn set_item_clamped(&self, item: ItemStack) { ... }
    fn save_throwable_item(&self, nbt: &mut NbtCompound) { ... }
    fn load_throwable_item(&self, nbt: BorrowedNbtCompoundView<'_, '_>) { ... }
}

Vanilla-shaped behavior shared by entities that extend ThrowableItemProjectile.

Required Methods

fn get_default_item(&self) -> ItemRef

Vanilla ThrowableItemProjectile.getDefaultItem.

fn set_item(&self, item: ItemStack)

Sets the rendered item stack (vanilla setItem, count clamped to 1).

The concrete entity stores this in its synced data layer.

fn get_item(&self) -> ItemStack

Returns the rendered item stack (vanilla getItem).

Provided Methods

fn set_item_clamped(&self, item: ItemStack)

Vanilla ThrowableItemProjectile.setItem count-clamping helper.

fn save_throwable_item(&self, nbt: &mut NbtCompound)

Saves the item stack (vanilla addAdditionalSaveData).

fn load_throwable_item(&self, nbt: BorrowedNbtCompoundView<'_, '_>)

Loads the item stack, falling back to the default item (vanilla readAdditionalSaveData).

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".