steel_core::entity

Trait Entity

pub trait Entity:
    EntityEventSource
    + Send
    + Sync {

    // Required methods
    fn base(&self) -> &EntityBase;
    fn entity_type(&self) -> EntityTypeRef;

    // Provided methods
    fn broadcast_to_player(&self, _player: &Player) -> bool { ... }
    fn id(&self) -> i32 { ... }
    fn uuid(&self) -> Uuid { ... }
    fn position(&self) -> DVec3 { ... }
    fn block_position(&self) -> BlockPos { ... }
    fn in_block_state(&self, world: &World) -> BlockStateId { ... }
    fn old_position(&self) -> DVec3 { ... }
    fn bounding_box(&self) -> WorldAabb { ... }
    fn is_free(&self, delta: DVec3) -> bool { ... }
    fn blocks_building(&self) -> bool { ... }
    fn is_pickable(&self) -> bool { ... }
    fn is_pushable(&self) -> bool { ... }
    fn is_pushed_by_fluid(&self) -> bool { ... }
    fn is_spectator(&self) -> bool { ... }
    fn can_be_collided_with(&self, _other: Option<&dyn Entity>) -> bool { ... }
    fn can_be_hit_by_projectile(&self) -> bool { ... }
    fn vehicle(&self) -> Option<SharedEntity> { ... }
    fn is_passenger(&self) -> bool { ... }
    fn stop_riding(&self) { ... }
    fn passengers(&self) -> Vec<SharedEntity>  { ... }
    fn count_player_passengers(&self) -> usize { ... }
    fn has_exactly_one_player_passenger(&self) -> bool { ... }
    fn first_passenger(&self) -> Option<SharedEntity> { ... }
    fn controlling_passenger(&self) -> Option<SharedEntity> { ... }
    fn has_controlling_passenger(&self) -> bool { ... }
    fn is_vehicle(&self) -> bool { ... }
    fn has_passenger(&self, passenger: &dyn Entity) -> bool { ... }
    fn passenger_index(&self, passenger: &dyn Entity) -> Option<usize> { ... }
    fn vehicle_attachment_point(&self, _vehicle: &dyn Entity) -> DVec3 { ... }
    fn passenger_attachment_point(&self, passenger: &dyn Entity) -> DVec3 { ... }
    fn passenger_riding_position(&self, passenger: &dyn Entity) -> DVec3 { ... }
    fn position_rider(&self, passenger: &dyn Entity) { ... }
    fn root_vehicle_id(&self) -> i32 { ... }
    fn root_vehicle(&self) -> Option<SharedEntity> { ... }
    fn is_passenger_of_same_vehicle(&self, other: &dyn Entity) -> bool { ... }
    fn has_indirect_passenger(&self, entity: &dyn Entity) -> bool { ... }
    fn can_collide_with(&self, other: &dyn Entity) -> bool { ... }
    fn push_impulse(&self, impulse: DVec3) { ... }
    fn push_entity(&self, entity: &dyn Entity) { ... }
    fn make_bounding_box_at(&self, position: DVec3) -> WorldAabb { ... }
    fn default_tick(&self) { ... }
    fn tick(&self) { ... }
    fn ride_tick(&self) { ... }
    fn base_tick(&self) { ... }
    fn check_below_world(&self) { ... }
    fn on_below_world(&self) { ... }
    fn apply_inside_block_effect(&self, effect_type: InsideBlockEffectType) { ... }
    fn level(&self) -> Option<Arc<World>> { ... }
    fn pack_dirty_entity_data(&self) -> Option<Vec<DataValue>> { ... }
    fn pack_all_entity_data(&self) -> Vec<DataValue> { ... }
    fn synced_data(&self) -> Option<&dyn EntitySyncedData> { ... }
    fn pack_syncable_attributes(&self) -> Vec<AttributeSnapshot> { ... }
    fn drain_dirty_syncable_attributes(&self) -> Vec<AttributeSnapshot> { ... }
    fn is_removed(&self) -> bool { ... }
    fn is_alive(&self) -> bool { ... }
    fn removal_reason(&self) -> Option<RemovalReason> { ... }
    fn set_removed(&self, reason: RemovalReason) { ... }
    fn set_level_callback(&self, callback: Arc<dyn EntityLevelCallback>) { ... }
    fn as_item_entity(self: Arc<Self>) -> Option<Arc<ItemEntity>> { ... }
    fn is_living_entity(&self) -> bool { ... }
    fn forces_fall_flying_velocity_sync(&self) -> bool { ... }
    fn uses_client_movement_packets(&self) -> bool { ... }
    fn is_server_driven_movement(&self) -> bool { ... }
    fn can_simulate_movement(&self) -> bool { ... }
    fn is_effective_ai(&self) -> bool { ... }
    fn is_suppressing_bounce(&self) -> bool { ... }
    fn is_stepping_carefully(&self) -> bool { ... }
    fn is_descending(&self) -> bool { ... }
    fn set_shared_shift_key_down(&self, shift_key_down: bool) { ... }
    fn set_shared_swimming(&self, swimming: bool) { ... }
    fn set_shared_sprinting(&self, sprinting: bool) { ... }
    fn set_shared_fall_flying(&self, fall_flying: bool) { ... }
    fn default_can_walk_on_powder_snow(&self) -> bool { ... }
    fn can_walk_on_powder_snow(&self) -> bool { ... }
    fn is_flying_vehicle(&self) -> bool { ... }
    fn on_climbable(&self) -> bool { ... }
    fn known_movement(&self) -> DVec3 { ... }
    fn known_speed(&self) -> DVec3 { ... }
    fn tick_count(&self) -> i32 { ... }
    fn advance_tick_count(&self) { ... }
    fn fall_sounds(&self) -> (SoundEventRef, SoundEventRef) { ... }
    fn rotation(&self) -> (f32, f32) { ... }
    fn set_rotation(&self, rotation: (f32, f32)) { ... }
    fn spawn_data(&self) -> i32 { ... }
    fn get_eye_height(&self) -> f64 { ... }
    fn get_fluid_jump_threshold(&self) -> f64 { ... }
    fn get_eye_y(&self) -> f64 { ... }
    fn calculate_view_vector(
        &self,
        pitch_degrees: f32,
        yaw_degrees: f32,
    ) -> DVec3 { ... }
    fn look_angle(&self) -> DVec3 { ... }
    fn velocity(&self) -> DVec3 { ... }
    fn set_velocity(&self, velocity: DVec3) { ... }
    fn needs_velocity_sync(&self) -> bool { ... }
    fn mark_velocity_sync(&self) { ... }
    fn clear_velocity_sync(&self) { ... }
    fn fall_distance(&self) -> f64 { ... }
    fn is_in_powder_snow(&self) -> bool { ... }
    fn was_in_powder_snow(&self) -> bool { ... }
    fn set_fall_distance(&self, fall_distance: f64) { ... }
    fn reset_fall_distance(&self) { ... }
    fn check_fall_distance_accumulation(&self) { ... }
    fn fire_freeze_state(&self) -> EntityFireFreezeState { ... }
    fn remaining_fire_ticks(&self) -> i32 { ... }
    fn set_remaining_fire_ticks(&self, remaining_fire_ticks: i32) { ... }
    fn ticks_frozen(&self) -> i32 { ... }
    fn set_ticks_frozen(&self, ticks_frozen: i32) { ... }
    fn fire_immune(&self) -> bool { ... }
    fn fire_immune_ticks(&self) -> i32 { ... }
    fn should_play_lava_hurt_sound(&self) -> bool { ... }
    fn lava_hurt(&self) { ... }
    fn remaining_fire_ticks_cap(&self) -> Option<i32> { ... }
    fn fire_ignite_extra_ticks(&self) -> i32 { ... }
    fn is_on_fire(&self) -> bool { ... }
    fn has_visual_fire(&self) -> bool { ... }
    fn is_freezing(&self) -> bool { ... }
    fn default_can_freeze(&self) -> bool { ... }
    fn can_freeze(&self) -> bool { ... }
    fn ticks_required_to_freeze(&self) -> i32 { ... }
    fn is_fully_frozen(&self) -> bool { ... }
    fn clear_freeze(&self) { ... }
    fn clear_fire(&self) { ... }
    fn ignite_for_ticks(&self, number_of_ticks: i32) { ... }
    fn sync_base_fire_freeze_entity_data(&self) { ... }
    fn is_in_water(&self) -> bool { ... }
    fn is_in_lava(&self) -> bool { ... }
    fn is_eye_in_water(&self) -> bool { ... }
    fn is_eye_in_lava(&self) -> bool { ... }
    fn is_under_water(&self) -> bool { ... }
    fn fluid_contact(&self) -> EntityFluidContact { ... }
    fn refresh_fluid_contact(&self) -> EntityFluidContact { ... }
    fn refresh_fluid_contact_for_base_tick(&self) -> EntityFluidContact { ... }
    fn scan_and_store_fluid_contact(
        &self,
        advance_eye_water_history: bool,
    ) -> EntityFluidContact { ... }
    fn apply_fluid_current_for_base_tick(
        &self,
        world: &Arc<World>,
        contact: EntityFluidContact,
    ) { ... }
    fn apply_fluid_current_impulse(&self, impulse: DVec3) { ... }
    fn is_fall_damage_immune(&self) -> bool { ... }
    fn cause_fall_damage(
        &self,
        fall_distance: f64,
        damage_modifier: f32,
        source: &DamageSource,
    ) -> bool { ... }
    fn on_ground(&self) -> bool { ... }
    fn horizontal_collision(&self) -> bool { ... }
    fn vertical_collision(&self) -> bool { ... }
    fn vertical_collision_below(&self) -> bool { ... }
    fn no_physics(&self) -> bool { ... }
    fn is_affected_by_blocks(&self) -> bool { ... }
    fn set_no_physics(&self, no_physics: bool) { ... }
    fn update_no_physics_from_current_collision(&self) { ... }
    fn move_towards_closest_space(&self, x: f64, y: f64, z: f64) { ... }
    fn default_make_stuck_in_block(
        &self,
        _state: BlockStateId,
        speed_multiplier: DVec3,
    ) { ... }
    fn make_stuck_in_block(&self, state: BlockStateId, speed_multiplier: DVec3) { ... }
    fn apply_effects_from_blocks(&self) { ... }
    fn apply_effects_from_blocks_for_last_movements(&self) { ... }
    fn set_on_ground(&self, on_ground: bool) { ... }
    fn set_on_ground_with_movement(
        &self,
        on_ground: bool,
        horizontal_collision: bool,
        movement: DVec3,
    ) { ... }
    fn default_apply_accepted_client_movement(
        &self,
        world: &Arc<World>,
        accepted: AcceptedClientMovement,
    ) -> Result<AcceptedClientMovementOutcome, EntityMoveError> { ... }
    fn apply_accepted_client_movement(
        &self,
        world: &Arc<World>,
        accepted: AcceptedClientMovement,
    ) -> Result<AcceptedClientMovementOutcome, EntityMoveError> { ... }
    fn apply_accepted_client_vehicle_movement(
        &self,
        world: &Arc<World>,
        accepted: AcceptedClientMovement,
    ) -> Result<AcceptedClientMovementOutcome, EntityMoveError> { ... }
    fn try_set_position(&self, pos: DVec3) -> Result<(), EntityMoveError> { ... }
    fn set_old_position_to_current(&self) { ... }
    fn set_old_position(&self, old_position: DVec3) { ... }
    fn remove_latest_movement_recording(&self) { ... }
    fn on_pos(&self, offset: f32) -> Option<BlockPos> { ... }
    fn block_pos_below_that_affects_movement(&self) -> Option<BlockPos> { ... }
    fn on_pos_legacy(&self) -> Option<BlockPos> { ... }
    fn block_speed_factor(&self) -> f32 { ... }
    fn block_jump_factor(&self) -> f32 { ... }
    fn pose(&self) -> EntityPose { ... }
    fn is_crouching(&self) -> bool { ... }
    fn is_swimming(&self) -> bool { ... }
    fn is_on_rails(&self) -> bool { ... }
    fn is_state_climbable(&self, state: BlockStateId) -> bool { ... }
    fn movement_emission(&self) -> EntityMovementEmission { ... }
    fn may_interact(&self, _world: &World, _pos: BlockPos) -> bool { ... }
    fn sound_source(&self) -> SoundSource { ... }
    fn swim_sound(&self) -> SoundEventRef { ... }
    fn is_silent(&self) -> bool { ... }
    fn broadcast_entity_event(&self, event: EntityStatus) { ... }
    fn play_sound(&self, sound: SoundEventRef, volume: f32, pitch: f32) { ... }
    fn play_entity_on_fire_extinguished_sound(&self) { ... }
    fn play_step_sound(&self, _pos: BlockPos, block_state: BlockStateId) { ... }
    fn play_block_step_sound(&self, block_state: BlockStateId) { ... }
    fn play_muffled_step_sound(&self, block_state: BlockStateId) { ... }
    fn play_combination_step_sounds(
        &self,
        primary_step_sound: BlockStateId,
        secondary_step_sound: BlockStateId,
    ) { ... }
    fn walking_step_sound(&self, pos: BlockPos, block_state: BlockStateId) { ... }
    fn play_amethyst_step_sound(&self) { ... }
    fn water_swim_sound(&self) { ... }
    fn play_swim_sound(&self, volume: f32) { ... }
    fn is_flapping(&self) -> bool { ... }
    fn on_flap(&self) { ... }
    fn process_flapping_movement(&self) { ... }
    fn next_step(&self) -> f32 { ... }
    fn apply_movement_emission_and_play_sound(
        &self,
        emission: EntityMovementEmission,
        clipped_movement: DVec3,
        effect_pos: BlockPos,
        effect_state: BlockStateId,
    ) { ... }
    fn apply_movement_side_effects_after_move(
        &self,
        world: &World,
        actual_movement: DVec3,
    ) { ... }
    fn vibration_and_sound_effects_from_block(
        &self,
        pos: BlockPos,
        block_state: BlockStateId,
        should_sound: bool,
        should_vibrate: bool,
        clipped_movement: DVec3,
    ) -> bool { ... }
    fn max_up_step(&self) -> f32 { ... }
    fn backs_off_from_edge(&self) -> bool { ... }
    fn get_default_gravity(&self) -> f64 { ... }
    fn is_no_gravity(&self) -> bool { ... }
    fn set_no_gravity(&self, no_gravity: bool) { ... }
    fn get_gravity(&self) -> f64 { ... }
    fn apply_gravity(&self) { ... }
    fn move_relative(&self, speed: f32, input: DVec3) { ... }
    fn move_without_physics(&self, delta: DVec3) -> Option<MoveResult> { ... }
    fn move_entity(
        &self,
        mover_type: MoverType,
        delta: DVec3,
    ) -> Option<MoveResult> { ... }
    fn apply_fall_damage_after_move(
        &self,
        result: &MoveResult,
        world: &Arc<World>,
    ) -> bool { ... }
    fn reset_fall_distance_on_resetting_clip(
        &self,
        world: &Arc<World>,
        movement: DVec3,
    ) { ... }
    fn do_check_fall_damage(
        &self,
        movement: DVec3,
        on_ground: bool,
        world: &Arc<World>,
    ) -> bool { ... }
    fn refresh_supporting_block_for_fall_damage(
        &self,
        movement: DVec3,
        on_ground: bool,
    ) { ... }
    fn check_fall_damage(
        &self,
        vertical_movement: f64,
        on_ground: bool,
        on_state: BlockStateId,
        pos: BlockPos,
        world: &Arc<World>,
    ) { ... }
    fn ground_contact_after_movement(
        &self,
        on_ground: bool,
        movement: Option<DVec3>,
    ) -> EntityGroundContact { ... }
    fn check_supporting_block(
        &self,
        on_ground: bool,
        movement: Option<DVec3>,
        world: &Arc<World>,
    ) -> EntityGroundContact { ... }
    fn spawn_at_location(
        &self,
        item: ItemStack,
        y_offset: f64,
    ) -> Option<Arc<ItemEntity>> { ... }
    fn save_additional(&self, _nbt: &mut NbtCompound) { ... }
    fn load_additional(&self, _nbt: &BaseNbtCompound<'_>) { ... }
    fn hurt(&self, source: &DamageSource, amount: f32) -> bool { ... }
    fn change_world(self: Arc<Self>, _teleport_transition: &TeleportTransition) { ... }

}

A trait for entities.

This trait provides the core functionality for entities. It’s based on Minecraft’s Entity class.

Using EntityBase

Entities expose EntityBase to get default implementations for common methods like id(), uuid(), position(), etc.

impl Entity for MyEntity {
    fn base(&self) -> &EntityBase { &self.base }
    fn entity_type(&self) -> EntityTypeRef { vanilla_entities::MY_ENTITY }
    fn bounding_box(&self) -> WorldAabb { /* ... */ }
    // All other common methods use defaults from EntityBase!
}

Required Methods

fn base(&self) -> &EntityBase

Returns a reference to the entity’s shared vanilla base fields.

fn entity_type(&self) -> EntityTypeRef

Gets the entity type containing tracking range, dimensions, etc.

Provided Methods

fn broadcast_to_player(&self, _player: &Player) -> bool

Returns whether this entity should be broadcast to the given player.

Mirrors vanilla Entity.broadcastToPlayer. Most entities are always broadcastable; players override this for spectator visibility rules.

fn id(&self) -> i32

Gets the entity’s unique network ID (session-local).

fn uuid(&self) -> Uuid

Gets the UUID of the entity (persistent identifier).

fn position(&self) -> DVec3

Gets the entity’s current position.

fn block_position(&self) -> BlockPos

Gets the entity’s current block position.

fn in_block_state(&self, world: &World) -> BlockStateId

Returns vanilla Entity.getInBlockState.

fn old_position(&self) -> DVec3

Gets the entity position used by vanilla movement traces.

fn bounding_box(&self) -> WorldAabb

Gets the entity’s bounding box for collision queries.

fn is_free(&self, delta: DVec3) -> bool

Returns vanilla Entity.isFree() for the current bounding box shifted by delta.

fn blocks_building(&self) -> bool

Returns whether this entity obstructs block placement.

Mirrors vanilla Entity.blocksBuilding. Base entities do not obstruct placement unless a concrete entity type opts in.

fn is_pickable(&self) -> bool

Returns whether this entity can be targeted by picking and interaction raycasts.

Mirrors vanilla Entity.isPickable. Base entities are not pickable unless a concrete entity type opts in.

fn is_pushable(&self) -> bool

Returns whether this entity participates in vanilla push separation.

Mirrors vanilla Entity.isPushable. Base entities are not pushable unless a concrete entity type opts in.

fn is_pushed_by_fluid(&self) -> bool

Returns whether vanilla fluid currents can push this entity.

fn is_spectator(&self) -> bool

Returns whether this entity is invisible to normal entity selectors.

Mirrors vanilla Entity.isSpectator. Base entities are never spectators; players override this from their game mode.

fn can_be_collided_with(&self, _other: Option<&dyn Entity>) -> bool

Returns whether other can collide with this entity.

Mirrors vanilla Entity.canBeCollidedWith. Base entities cannot be collided with unless a concrete entity type opts in.

fn can_be_hit_by_projectile(&self) -> bool

Returns whether projectile collision may interact with this entity.

Mirrors vanilla Entity.canBeHitByProjectile.

fn vehicle(&self) -> Option<SharedEntity>

Gets the vehicle this entity is riding, if present.

Mirrors vanilla Entity.getVehicle.

fn is_passenger(&self) -> bool

Returns whether this entity is riding another entity.

Mirrors vanilla Entity.isPassenger.

fn stop_riding(&self)

Stops riding the current vehicle, if any.

Mirrors vanilla Entity.stopRiding.

fn passengers(&self) -> Vec<SharedEntity>

Gets this entity’s direct passengers.

Mirrors vanilla Entity.getPassengers.

fn count_player_passengers(&self) -> usize

Counts indirect player passengers.

Mirrors vanilla Entity.countPlayerPassengers.

fn has_exactly_one_player_passenger(&self) -> bool

Returns whether this entity has exactly one indirect player passenger.

Mirrors vanilla Entity.hasExactlyOnePlayerPassenger.

fn first_passenger(&self) -> Option<SharedEntity>

Gets this entity’s first direct passenger.

Mirrors vanilla Entity.getFirstPassenger.

fn controlling_passenger(&self) -> Option<SharedEntity>

Returns the living passenger currently controlling this entity, if any.

Mirrors vanilla Entity.getControllingPassenger. Base entities have no controller; controllable vehicles override this based on their own rules.

fn has_controlling_passenger(&self) -> bool

Returns whether this entity currently has a controlling passenger.

Mirrors vanilla Entity.hasControllingPassenger.

fn is_vehicle(&self) -> bool

Returns whether this entity has any direct passengers.

Mirrors vanilla Entity.isVehicle.

fn has_passenger(&self, passenger: &dyn Entity) -> bool

Returns whether passenger is a direct passenger of this entity.

Mirrors vanilla Entity.hasPassenger(Entity).

fn passenger_index(&self, passenger: &dyn Entity) -> Option<usize>

Returns the current direct passenger index for attachment lookup.

fn vehicle_attachment_point(&self, _vehicle: &dyn Entity) -> DVec3

Returns this passenger’s vehicle attachment point.

Mirrors vanilla Entity.getVehicleAttachmentPoint.

fn passenger_attachment_point(&self, passenger: &dyn Entity) -> DVec3

Returns this vehicle’s passenger attachment point.

Mirrors vanilla Entity.getPassengerAttachmentPoint for the base entity class.

fn passenger_riding_position(&self, passenger: &dyn Entity) -> DVec3

Returns the world position where passenger should ride this vehicle.

Mirrors vanilla Entity.getPassengerRidingPosition.

fn position_rider(&self, passenger: &dyn Entity)

Repositions a direct passenger from this vehicle’s attachment point.

Mirrors vanilla Entity.positionRider.

fn root_vehicle_id(&self) -> i32

Returns this entity’s root vehicle ID, or this entity’s ID when it is not riding.

Mirrors vanilla Entity.getRootVehicle using session IDs for object identity.

fn root_vehicle(&self) -> Option<SharedEntity>

Returns this entity’s root vehicle, if this entity is riding one.

Mirrors vanilla Entity.getRootVehicle.

fn is_passenger_of_same_vehicle(&self, other: &dyn Entity) -> bool

Returns whether this entity and other share the same root vehicle.

Mirrors vanilla Entity.isPassengerOfSameVehicle.

fn has_indirect_passenger(&self, entity: &dyn Entity) -> bool

Returns whether entity is an indirect passenger of this entity.

Mirrors vanilla Entity.hasIndirectPassenger.

fn can_collide_with(&self, other: &dyn Entity) -> bool

Returns whether this entity can collide with other.

Mirrors vanilla Entity.canCollideWith.

fn push_impulse(&self, impulse: DVec3)

Adds an impulse to this entity’s velocity and marks velocity for sync.

Mirrors vanilla Entity.push(double, double, double).

fn push_entity(&self, entity: &dyn Entity)

Applies vanilla entity-to-entity push separation.

Mirrors vanilla Entity.push(Entity).

fn make_bounding_box_at(&self, position: DVec3) -> WorldAabb

Builds this entity’s default bounding box at position.

fn default_tick(&self)

Default vanilla Entity.tick() behavior.

Concrete entity ticks that mirror vanilla super.tick() should call this rather than calling Self::base_tick directly.

fn tick(&self)

Called every game tick when the entity is in a ticked chunk.

Use self.level() to access the world for physics, block queries, etc. The caller handles post-tick dirty data sync.

Steel keeps the fallback empty because many vanilla subclasses override tick without calling super.tick().

fn ride_tick(&self)

Called every game tick while this entity is riding another entity.

Mirrors vanilla Entity.rideTick.

fn base_tick(&self)

Runs the vanilla base-tick physics pieces Steel currently implements.

This intentionally stays separate from tick() because several vanilla subclasses override tick without calling super.tick().

fn check_below_world(&self)

Applies vanilla below-world handling.

fn on_below_world(&self)

Runs entity-specific behavior after falling below the world.

fn apply_inside_block_effect(&self, effect_type: InsideBlockEffectType)

Applies an inside-block effect queued by vanilla’s step-based collector.

fn level(&self) -> Option<Arc<World>>

Gets the world this entity is in.

Returns None if the entity is not in a world or the world was dropped. Mirrors vanilla’s Entity.level().

fn pack_dirty_entity_data(&self) -> Option<Vec<DataValue>>

Packs dirty entity data for network synchronization.

Returns Some(values) if there are dirty values to sync, None otherwise. Clears the dirty flags after packing.

fn pack_all_entity_data(&self) -> Vec<DataValue>

Packs all non-default entity data for initial spawn.

Used when sending entity data to a player who just started tracking this entity.

fn synced_data(&self) -> Option<&dyn EntitySyncedData>

Returns the synchronized entity-data container for entities with vanilla data accessors.

fn pack_syncable_attributes(&self) -> Vec<AttributeSnapshot>

Packs syncable attributes for initial spawn pairing.

Mirrors vanilla ServerEntity.sendPairingData, which sends all syncable living attributes after the add-entity and metadata packets.

fn drain_dirty_syncable_attributes(&self) -> Vec<AttributeSnapshot>

Drains syncable dirty attributes for per-tick tracking updates.

Mirrors vanilla ServerEntity.sendDirtyEntityData, which sends dirty living attributes after dirty entity data.

fn is_removed(&self) -> bool

Returns true if the entity has been marked for removal.

fn is_alive(&self) -> bool

Returns whether this entity is alive for vanilla generic entity checks.

fn removal_reason(&self) -> Option<RemovalReason>

Returns why this entity was removed, if it has been removed.

fn set_removed(&self, reason: RemovalReason)

Marks the entity as removed with the given reason.

fn set_level_callback(&self, callback: Arc<dyn EntityLevelCallback>)

Sets the level callback for lifecycle events (movement, removal).

fn as_item_entity(self: Arc<Self>) -> Option<Arc<ItemEntity>>

Gets the entity as an ItemEntity if it is one.

fn is_living_entity(&self) -> bool

Returns true for entities that implement vanilla living-entity behavior.

fn forces_fall_flying_velocity_sync(&self) -> bool

Returns true when vanilla ServerEntity should force velocity sync for fall flying.

fn uses_client_movement_packets(&self) -> bool

Returns true when movement is driven by serverbound movement packets.

fn is_server_driven_movement(&self) -> bool

Returns true when normal server ticks drive this entity’s movement.

fn can_simulate_movement(&self) -> bool

Returns true when vanilla allows this side to apply movement simulation side effects.

fn is_effective_ai(&self) -> bool

Returns true when vanilla allows this side to run entity AI/travel logic.

fn is_suppressing_bounce(&self) -> bool

Returns true when vanilla landing bounce should be suppressed.

fn is_stepping_carefully(&self) -> bool

Returns true when vanilla block step-on hooks should treat this entity as careful.

fn is_descending(&self) -> bool

Returns true when vanilla collision context should treat the entity as descending.

fn set_shared_shift_key_down(&self, shift_key_down: bool)

Sets the vanilla shift-key-down shared flag.

fn set_shared_swimming(&self, swimming: bool)

Sets the vanilla swimming shared flag.

fn set_shared_sprinting(&self, sprinting: bool)

Sets the vanilla sprinting shared flag.

fn set_shared_fall_flying(&self, fall_flying: bool)

Sets the vanilla fall-flying shared flag.

fn default_can_walk_on_powder_snow(&self) -> bool

Returns vanilla PowderSnowBlock.canEntityWalkOnPowderSnow.

fn can_walk_on_powder_snow(&self) -> bool

Returns whether this entity can walk on powder snow.

fn is_flying_vehicle(&self) -> bool

Returns whether vanilla excludes this vehicle from floating kicks.

fn on_climbable(&self) -> bool

Returns true if vanilla rules consider this entity to be on a climbable block.

fn known_movement(&self) -> DVec3

Returns the movement vector vanilla exposes for block-contact logic.

fn known_speed(&self) -> DVec3

Returns the base-tick displacement vanilla exposes as getKnownSpeed.

fn tick_count(&self) -> i32

Returns vanilla Entity.tickCount.

fn advance_tick_count(&self)

Advances vanilla Entity.tickCount.

fn fall_sounds(&self) -> (SoundEventRef, SoundEventRef)

Returns vanilla small and big fall sounds for this entity.

fn rotation(&self) -> (f32, f32)

Gets the entity’s rotation as (yaw, pitch) in degrees.

Yaw is horizontal rotation (0-360), pitch is vertical (-90 to 90).

fn set_rotation(&self, rotation: (f32, f32))

Sets the entity’s rotation as (yaw, pitch) in degrees.

fn spawn_data(&self) -> i32

Extra spawn-packet data used by vanilla for entity-specific construction.

fn get_eye_height(&self) -> f64

Gets the eye height for this entity.

Default implementation returns the eye height from the entity type dimensions. Override for entities with pose-dependent eye heights (e.g., players).

fn get_fluid_jump_threshold(&self) -> f64

Returns vanilla Entity.getFluidJumpThreshold().

fn get_eye_y(&self) -> f64

Gets the Y coordinate of the entity’s eyes.

Equivalent to vanilla’s Entity.getEyeY().

fn calculate_view_vector(&self, pitch_degrees: f32, yaw_degrees: f32) -> DVec3

Calculates vanilla Entity.calculateViewVector().

fn look_angle(&self) -> DVec3

Returns vanilla Entity.getLookAngle().

fn velocity(&self) -> DVec3

Gets the entity’s velocity in blocks per tick.

fn set_velocity(&self, velocity: DVec3)

Sets the entity’s velocity.

fn needs_velocity_sync(&self) -> bool

Returns true when vanilla ServerEntity should consider sending velocity.

fn mark_velocity_sync(&self)

Marks velocity for vanilla ServerEntity synchronization.

fn clear_velocity_sync(&self)

Clears the vanilla velocity sync marker after send processing.

fn fall_distance(&self) -> f64

Returns accumulated vanilla fall distance.

fn is_in_powder_snow(&self) -> bool

Returns whether this entity is currently inside powder snow.

fn was_in_powder_snow(&self) -> bool

Returns whether this entity was inside powder snow during the previous base tick.

fn set_fall_distance(&self, fall_distance: f64)

Sets accumulated vanilla fall distance.

fn reset_fall_distance(&self)

Resets accumulated vanilla fall distance.

fn check_fall_distance_accumulation(&self)

Mirrors vanilla Entity.checkFallDistanceAccumulation().

fn fire_freeze_state(&self) -> EntityFireFreezeState

Returns the current vanilla fire/freeze state.

fn remaining_fire_ticks(&self) -> i32

Returns vanilla remainingFireTicks.

fn set_remaining_fire_ticks(&self, remaining_fire_ticks: i32)

Sets vanilla remainingFireTicks.

fn ticks_frozen(&self) -> i32

Returns synchronized vanilla TicksFrozen.

fn set_ticks_frozen(&self, ticks_frozen: i32)

Sets synchronized vanilla TicksFrozen.

fn fire_immune(&self) -> bool

Returns whether this entity is immune to fire effects and fire damage.

fn fire_immune_ticks(&self) -> i32

Returns vanilla fire immunity cooldown ticks after not being ignited.

fn should_play_lava_hurt_sound(&self) -> bool

Returns whether vanilla should play this entity’s lava hurt sound.

fn lava_hurt(&self)

Applies vanilla lava-contact damage after lava ignition effects.

fn remaining_fire_ticks_cap(&self) -> Option<i32>

Maximum vanilla remainingFireTicks this entity can store.

fn fire_ignite_extra_ticks(&self) -> i32

Returns extra ticks added by fire-block ignition before 8-second ignition.

fn is_on_fire(&self) -> bool

Returns whether the entity is on fire on the server.

fn has_visual_fire(&self) -> bool

Returns vanilla hasVisualFire.

fn is_freezing(&self) -> bool

Returns whether the entity has any frozen ticks.

fn default_can_freeze(&self) -> bool

Returns vanilla Entity.canFreeze() without living-equipment overrides.

fn can_freeze(&self) -> bool

Returns whether this entity may accumulate frozen ticks.

fn ticks_required_to_freeze(&self) -> i32

Returns vanilla getTicksRequiredToFreeze.

fn is_fully_frozen(&self) -> bool

Returns whether this entity has reached full-freeze duration.

fn clear_freeze(&self)

Clears accumulated freezing.

fn clear_fire(&self)

Clears fire without resetting the vanilla fire immunity cooldown.

fn ignite_for_ticks(&self, number_of_ticks: i32)

Ignites this entity for a vanilla tick duration.

fn sync_base_fire_freeze_entity_data(&self)

Projects base fire/freeze state into generated synced entity data.

fn is_in_water(&self) -> bool

Returns true if this entity is currently touching water.

fn is_in_lava(&self) -> bool

Returns true if this entity is currently touching lava.

fn is_eye_in_water(&self) -> bool

Returns true if this entity’s eyes are currently inside water.

fn is_eye_in_lava(&self) -> bool

Returns true if this entity’s eyes are currently inside lava.

fn is_under_water(&self) -> bool

Returns vanilla underwater state.

fn fluid_contact(&self) -> EntityFluidContact

Returns cached fluid contact from the last entity fluid refresh.

fn refresh_fluid_contact(&self) -> EntityFluidContact

Refreshes cached fluid contact from this entity’s current bounding box.

fn refresh_fluid_contact_for_base_tick(&self) -> EntityFluidContact

Refreshes cached fluid contact with vanilla base-tick eye-water history.

fn scan_and_store_fluid_contact( &self, advance_eye_water_history: bool, ) -> EntityFluidContact

Scans current fluid contact and stores it on the entity base.

fn apply_fluid_current_for_base_tick( &self, world: &Arc<World>, contact: EntityFluidContact, )

Applies vanilla water/lava current impulses from the base-tick fluid scan.

fn apply_fluid_current_impulse(&self, impulse: DVec3)

Applies a non-zero fluid current impulse and marks velocity sync.

fn is_fall_damage_immune(&self) -> bool

Returns true if this entity type ignores vanilla fall damage.

fn cause_fall_damage( &self, fall_distance: f64, damage_modifier: f32, source: &DamageSource, ) -> bool

Applies vanilla fall damage. Base entities only propagate to passengers.

fn on_ground(&self) -> bool

Returns true if the entity is on the ground.

fn horizontal_collision(&self) -> bool

Returns true if the last movement was clipped horizontally.

fn vertical_collision(&self) -> bool

Returns true if the last movement was clipped vertically.

fn vertical_collision_below(&self) -> bool

Returns true if the last vertical collision was below the entity.

fn no_physics(&self) -> bool

Returns true when movement bypasses collision physics.

fn is_affected_by_blocks(&self) -> bool

Returns true when vanilla block-contact effects may run for this entity.

fn set_no_physics(&self, no_physics: bool)

Sets whether this entity bypasses collision physics.

fn update_no_physics_from_current_collision(&self)

Updates item-style noPhysics from the entity’s current collision state.

fn move_towards_closest_space(&self, x: f64, y: f64, z: f64)

Nudges velocity toward the closest non-full collision block.

fn default_make_stuck_in_block( &self, _state: BlockStateId, speed_multiplier: DVec3, )

Default vanilla stuck-in-block movement for the next movement pass.

fn make_stuck_in_block(&self, state: BlockStateId, speed_multiplier: DVec3)

Applies vanilla stuck-in-block movement for the next movement pass.

fn apply_effects_from_blocks(&self)

Applies current block-contact effects to this entity.

Mirrors the shared ownership boundary of vanilla Entity.applyEffectsFromBlocks.

fn apply_effects_from_blocks_for_last_movements(&self)

Replays the last finalized block-contact movement list.

fn set_on_ground(&self, on_ground: bool)

Sets whether the entity is on the ground.

fn set_on_ground_with_movement( &self, on_ground: bool, horizontal_collision: bool, movement: DVec3, )

Sets ground and horizontal collision flags from accepted movement.

fn default_apply_accepted_client_movement( &self, world: &Arc<World>, accepted: AcceptedClientMovement, ) -> Result<AcceptedClientMovementOutcome, EntityMoveError>

Default final state application for accepted client-authored movement.

Mirrors the shared tail of vanilla player and controlled-vehicle movement handling after rollback/collision validation has accepted the target.

fn apply_accepted_client_movement( &self, world: &Arc<World>, accepted: AcceptedClientMovement, ) -> Result<AcceptedClientMovementOutcome, EntityMoveError>

Applies final state accepted from a client-authored movement packet.

fn apply_accepted_client_vehicle_movement( &self, world: &Arc<World>, accepted: AcceptedClientMovement, ) -> Result<AcceptedClientMovementOutcome, EntityMoveError>

Applies final state accepted from a controlled-vehicle movement packet.

fn try_set_position(&self, pos: DVec3) -> Result<(), EntityMoveError>

Attempts to set the entity’s position through world lifecycle validation.

fn set_old_position_to_current(&self)

Sets the vanilla movement-trace old position to the current position.

fn set_old_position(&self, old_position: DVec3)

Sets the vanilla movement-trace old position explicitly.

fn remove_latest_movement_recording(&self)

Removes the latest movement segment recorded this tick.

fn on_pos(&self, offset: f32) -> Option<BlockPos>

Returns the block position this entity is standing on.

fn block_pos_below_that_affects_movement(&self) -> Option<BlockPos>

Returns the block position used for movement-affecting block properties.

fn on_pos_legacy(&self) -> Option<BlockPos>

Returns vanilla getOnPosLegacy(), used by fall/step block hooks.

fn block_speed_factor(&self) -> f32

Returns the vanilla block speed factor applied after movement.

fn block_jump_factor(&self) -> f32

Returns vanilla Entity.getBlockJumpFactor().

fn pose(&self) -> EntityPose

Returns this entity’s physical pose.

fn is_crouching(&self) -> bool

Returns whether vanilla currently considers this entity crouching.

fn is_swimming(&self) -> bool

Returns whether vanilla currently considers this entity swimming.

fn is_on_rails(&self) -> bool

Returns whether this entity is on rails.

fn is_state_climbable(&self, state: BlockStateId) -> bool

Returns whether a block state is climbable for base movement effects.

fn movement_emission(&self) -> EntityMovementEmission

Returns vanilla movement side effects emitted by this entity.

fn may_interact(&self, _world: &World, _pos: BlockPos) -> bool

Returns whether this entity may modify the world at a position.

Vanilla Entity.mayInteract defaults to true; player-like entities can apply world permission checks through overrides.

fn sound_source(&self) -> SoundSource

Returns this entity’s vanilla sound source category.

fn swim_sound(&self) -> SoundEventRef

Returns this entity’s vanilla swim sound.

fn is_silent(&self) -> bool

Returns whether sounds from this entity are suppressed.

fn broadcast_entity_event(&self, event: EntityStatus)

Broadcasts a vanilla entity event/status packet near this entity.

fn play_sound(&self, sound: SoundEventRef, volume: f32, pitch: f32)

Plays an entity sound at the entity’s exact position.

fn play_entity_on_fire_extinguished_sound(&self)

Plays vanilla’s extinguished-on-fire entity sound.

fn play_step_sound(&self, _pos: BlockPos, block_state: BlockStateId)

Plays the base vanilla step sound for a block.

fn play_block_step_sound(&self, block_state: BlockStateId)

Plays a vanilla block step sound at this entity’s current position.

fn play_muffled_step_sound(&self, block_state: BlockStateId)

Plays vanilla’s muffled secondary step sound.

fn play_combination_step_sounds( &self, primary_step_sound: BlockStateId, secondary_step_sound: BlockStateId, )

Plays vanilla’s combination primary and secondary step sounds.

fn walking_step_sound(&self, pos: BlockPos, block_state: BlockStateId)

Plays vanilla walking step sounds, including amethyst chimes.

fn play_amethyst_step_sound(&self)

Plays vanilla amethyst step chime when its cooldown permits it.

fn water_swim_sound(&self)

Plays vanilla swim sound from movement emission.

fn play_swim_sound(&self, volume: f32)

Plays this entity’s swim sound at the given volume.

fn is_flapping(&self) -> bool

Returns whether the entity is currently flapping.

fn on_flap(&self)

Runs entity-specific flap side effects.

fn process_flapping_movement(&self)

Processes vanilla flap movement side effects.

fn next_step(&self) -> f32

Returns the next step threshold after movement side effects are produced.

fn apply_movement_emission_and_play_sound( &self, emission: EntityMovementEmission, clipped_movement: DVec3, effect_pos: BlockPos, effect_state: BlockStateId, )

Applies vanilla movement sounds and game events after a completed move.

fn apply_movement_side_effects_after_move( &self, world: &World, actual_movement: DVec3, )

Applies movement side effects after vanilla collision and landing updates.

fn vibration_and_sound_effects_from_block( &self, pos: BlockPos, block_state: BlockStateId, should_sound: bool, should_vibrate: bool, clipped_movement: DVec3, ) -> bool

Emits step side effects from a candidate movement-effect block.

fn max_up_step(&self) -> f32

Maximum height this entity can step up during normal movement.

fn backs_off_from_edge(&self) -> bool

Whether movement should apply player-style sneak edge prevention.

fn get_default_gravity(&self) -> f64

Gets the default gravity for this entity type.

Override this to specify entity-specific gravity. Vanilla values: ItemEntity = 0.04, LivingEntity = 0.08

fn is_no_gravity(&self) -> bool

Returns true if gravity is disabled for this entity.

fn set_no_gravity(&self, no_gravity: bool)

Sets the shared vanilla NoGravity flag.

fn get_gravity(&self) -> f64

Gets the current gravity value.

Returns 0 if no_gravity is set, otherwise returns get_default_gravity().

fn apply_gravity(&self)

Applies gravity to the entity’s velocity.

Mirrors vanilla’s Entity.applyGravity().

fn move_relative(&self, speed: f32, input: DVec3)

Applies vanilla Entity.moveRelative().

fn move_without_physics(&self, delta: DVec3) -> Option<MoveResult>

Moves the entity without collision physics.

fn move_entity(&self, mover_type: MoverType, delta: DVec3) -> Option<MoveResult>

Moves the entity with collision detection.

Mirrors vanilla’s Entity.move(MoverType, Vec3). Updates position, on_ground, velocity (on collision), and returns collision info.

fn apply_fall_damage_after_move( &self, result: &MoveResult, world: &Arc<World>, ) -> bool

Applies vanilla fall-distance bookkeeping after accepted movement.

fn reset_fall_distance_on_resetting_clip( &self, world: &Arc<World>, movement: DVec3, )

Resets fall distance when vanilla’s fall-damage-resetting clip hits.

fn do_check_fall_damage( &self, movement: DVec3, on_ground: bool, world: &Arc<World>, ) -> bool

Mirrors vanilla Entity.doCheckFallDamage.

Callers update on-ground/supporting-block state before this method.

fn refresh_supporting_block_for_fall_damage( &self, movement: DVec3, on_ground: bool, )

Refreshes vanilla supporting-block state before fall-damage side effects.

fn check_fall_damage( &self, vertical_movement: f64, on_ground: bool, on_state: BlockStateId, pos: BlockPos, world: &Arc<World>, )

Mirrors vanilla Entity.checkFallDamage.

fn ground_contact_after_movement( &self, on_ground: bool, movement: Option<DVec3>, ) -> EntityGroundContact

Computes vanilla support state for an on-ground update.

fn check_supporting_block( &self, on_ground: bool, movement: Option<DVec3>, world: &Arc<World>, ) -> EntityGroundContact

Mirrors vanilla Entity.checkSupportingBlock.

fn spawn_at_location( &self, item: ItemStack, y_offset: f64, ) -> Option<Arc<ItemEntity>>

Spawns an item at this entity’s location.

Mirrors vanilla’s Entity.spawnAtLocation(). The item spawns at the entity’s position with the given Y offset and has a default pickup delay.

Returns None if the item stack is empty or the entity has no world.

fn save_additional(&self, _nbt: &mut NbtCompound)

Saves type-specific entity data to NBT.

Called during chunk serialization. Implementors should save all data needed to restore entity state on load. Base fields (pos, motion, rotation, uuid, on_ground) are handled by the serialization layer.

Mirrors vanilla’s Entity.addAdditionalSaveData().

fn load_additional(&self, _nbt: &BaseNbtCompound<'_>)

Loads type-specific entity data from NBT.

Called after entity creation during chunk deserialization. Base fields are already restored; this handles type-specific data.

Mirrors vanilla’s Entity.readAdditionalSaveData().

fn hurt(&self, source: &DamageSource, amount: f32) -> bool

Applies damage to this entity.

Vanilla: Entity.hurtServer() — overridden by LivingEntity (complex armor/effects/invulnerability logic) and ItemEntity (health decrement and discard). Default returns false (entity ignores damage).

fn change_world(self: Arc<Self>, _teleport_transition: &TeleportTransition)

Teleports an entity from one loaded world to another.

The default implementation logs a warning — non-player entity teleportation is not yet implemented. Override in entity types that support it.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".