steel_core::fluid::fluid_behavior

Trait FluidBehavior

pub trait FluidBehavior: Send + Sync {

    // Required methods
    fn fluid_type(&self) -> FluidRef;
    fn tick_delay(&self, world: &Arc<World>) -> i32;
    fn drop_off(&self, world: &Arc<World>) -> u8;
    fn slope_find_distance(&self, world: &Arc<World>) -> u8;
    fn tick(&self, world: &Arc<World>, pos: BlockPos);
    fn spread(&self, world: &Arc<World>, pos: BlockPos, fluid_state: FluidState);
    fn can_be_replaced_with(
        &self,
        fluid_state: FluidState,
        world: &Arc<World>,
        pos: BlockPos,
        other_fluid: FluidRef,
        direction: Direction,
    ) -> bool;

    // Provided methods
    fn is_same(&self, other: FluidRef) -> bool { ... }
    fn before_destroying_block(
        &self,
        _world: &Arc<World>,
        _pos: BlockPos,
        _replaced: BlockStateId,
    ) { ... }
    fn animate_tick(
        &self,
        world: &Arc<World>,
        pos: BlockPos,
        fluid_state: FluidState,
    ) { ... }
    fn can_convert_to_source(&self, _world: &Arc<World>) -> bool { ... }
    fn entity_inside(
        &self,
        world: &Arc<World>,
        pos: BlockPos,
        entity: &dyn Entity,
        effect_collector: &mut InsideBlockEffectCollector,
    ) { ... }
    fn explosion_resistance(&self) -> f32 { ... }
    fn is_randomly_ticking(&self) -> bool { ... }
    fn random_tick(&self, world: &Arc<World>, pos: BlockPos) { ... }
    fn get_spread_delay(
        &self,
        world: &Arc<World>,
        _pos: BlockPos,
        old_state: FluidState,
        new_state: FluidState,
    ) -> i32 { ... }
    fn get_flow(
        &self,
        world: &Arc<World>,
        pos: BlockPos,
        fluid_state: FluidState,
    ) -> DVec3 { ... }

}

Trait for fluid behavior implementations. Conceptual equivalent of Minecraft’s Fluid class.

Required Methods

fn fluid_type(&self) -> FluidRef

Gets the fluid type for this behavior.

fn tick_delay(&self, world: &Arc<World>) -> i32

Gets the number of ticks between fluid updates.

fn drop_off(&self, world: &Arc<World>) -> u8

Gets the amount of fluid level drop per horizontal block. Takes world because some fluids (lava) differ by dimension.

fn slope_find_distance(&self, world: &Arc<World>) -> u8

Gets the slope-search distance for horizontal spread. Takes world because some fluids (lava) differ by dimension.

fn tick(&self, world: &Arc<World>, pos: BlockPos)

Called every tick for fluid blocks.

fn spread(&self, world: &Arc<World>, pos: BlockPos, fluid_state: FluidState)

Called to calculate fluid spreading each tick.

fn can_be_replaced_with( &self, fluid_state: FluidState, world: &Arc<World>, pos: BlockPos, other_fluid: FluidRef, direction: Direction, ) -> bool

Checks if this fluid can be replaced by another fluid. This is used to determine if a fluid can flow into a block occupied by another fluid.

Provided Methods

fn is_same(&self, other: FluidRef) -> bool

Checks if this fluid is the same type as another fluid ref.

Used to determine if fluids can flow into each other.

Override required for any fluid that has both a source and a flowing variant

fn before_destroying_block( &self, _world: &Arc<World>, _pos: BlockPos, _replaced: BlockStateId, )

Called before a block is destroyed by this fluid.

fn animate_tick( &self, world: &Arc<World>, pos: BlockPos, fluid_state: FluidState, )

Called at tick time to play ambient animations (sounds, particles).

fn can_convert_to_source(&self, _world: &Arc<World>) -> bool

Checks if this fluid can convert to a source block at the given position.

fn entity_inside( &self, world: &Arc<World>, pos: BlockPos, entity: &dyn Entity, effect_collector: &mut InsideBlockEffectCollector, )

Called when an entity is inside this fluid.

fn explosion_resistance(&self) -> f32

Gets the explosion resistance of this fluid.

fn is_randomly_ticking(&self) -> bool

Returns whether this fluid should receive random ticks. Vanilla: only lava returns true (for fire spread).

fn random_tick(&self, world: &Arc<World>, pos: BlockPos)

Called on random tick for this fluid’s block. Used for lava fire spread.

fn get_spread_delay( &self, world: &Arc<World>, _pos: BlockPos, old_state: FluidState, new_state: FluidState, ) -> i32

Returns the tick delay to use when scheduling a newly-spread block, taking into account the old and new fluid states.

fn get_flow( &self, world: &Arc<World>, pos: BlockPos, fluid_state: FluidState, ) -> DVec3

Returns this fluid state’s vanilla flow vector at a position.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".