steel_core::inventory

Module container

Container trait for anything that holds items.

Containers are the base abstraction for anything that can hold items, including player inventories, chests, barrels, furnaces, etc.

Constants

DEFAULT_DISTANCE_BUFFER

Default distance buffer for container interaction range checks.

Traits

Container

Something that contains items. I also use container interchangeably with inventory as they mean approximately the same thing. But inventory could also refer to the player’s inventory. Example: PlayerInventory, Chest, Temporary Crafting Table

Functions

calculate_redstone_signal_from_container

Calculates the redstone comparator signal strength (0-15) from a container.

with_indices

Returns mutable references to N disjoint slots in a container.