steel_core::inventory::inventory_menu

Struct InventoryMenu

pub struct InventoryMenu {
    behavior: MenuBehavior,
    crafting_container: SyncCraftingContainer,
    result_container: SyncResultContainer,
}

The player’s inventory menu. This is always open when no other menu is open.

Fields
behavior: MenuBehaviorcrafting_container: SyncCraftingContainer

The 2x2 crafting grid container.

result_container: SyncResultContainer

The crafting result container.

Implementations

impl InventoryMenu

pub const CONTAINER_ID: u8 = 0

Container ID for the player inventory (always 0).

pub fn new(inventory: SyncPlayerInv) -> Self

Creates a new inventory menu for a player.

The inventory container should contain:

Slots 0-35: Main inventory (hotbar 0-8, main 9-35)

Slots 36-39: Armor (feet, legs, chest, head)

Slot 40: Offhand

pub fn update_crafting_result(&self)

Updates the crafting result based on the current grid contents. Should be called whenever a crafting grid slot changes.

pub const fn crafting_container(&self) -> &SyncCraftingContainer

Returns a reference to the crafting container.

pub const fn result_container(&self) -> &SyncResultContainer

Returns a reference to the result container.

pub fn crafting_container_ref(&self) -> ContainerRef

Returns a ContainerRef for the crafting container.

pub fn result_container_ref(&self) -> ContainerRef

Returns a ContainerRef for the result container.

pub fn is_hotbar_slot(slot: usize) -> bool

Returns true if the given slot index is in the hotbar or offhand. Based on Java’s InventoryMenu::isHotbarSlot.

pub fn is_armor_slot(slot: usize) -> bool

Returns true if the given slot index is an armor slot.

pub fn is_inventory_slot(slot: usize) -> bool

Returns true if the given slot index is in the main inventory.

fn move_between_inventory_and_hotbar( &self, guard: &mut ContainerLockGuard, slot_index: usize, stack: &mut ItemStack, ) -> bool

Helper method to move items between inventory and hotbar.