steel_core::inventory::menu

Trait SectionSource

pub trait SectionSource {
    // Required method
    fn take(self, count: usize) -> (ContainerRef, Range<usize>);
}

A supplier for ranges of slots in a ContainerRef. Allowing either making the whole Container a Section or splitting one Container into multiple Sections.

Required Methods

fn take(self, count: usize) -> (ContainerRef, Range<usize>)

Consumes the next count slots of the slot range.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".