steel_core::inventory

Module prelude

The most common imports needed when working with inventories and menus

Re-exports

pub use crate::inventory::click::Click;

pub use crate::inventory::click::ClickOutcome;

pub use crate::inventory::click::DragKind;

pub use crate::inventory::click::MouseButton;

pub use crate::inventory::click::QuickCraft;

pub use crate::inventory::click::SwapTarget;

pub use crate::inventory::container::Container;

pub use crate::inventory::container::SimpleContainer;

pub use crate::inventory::equipment::EntityEquipment;

pub use crate::inventory::equipment::OwnedEntityEquipment;

pub use crate::inventory::lock::ContainerId;

pub use crate::inventory::lock::ContainerLockGuard;

pub use crate::inventory::lock::ContainerRef;

pub use crate::inventory::menu::ContainerSlots;

pub use crate::inventory::menu::DataSlot;

pub use crate::inventory::menu::FakeResultRemainderPolicy;

pub use crate::inventory::menu::FillDirection;

pub use crate::inventory::menu::GridPlacer;

pub use crate::inventory::menu::Menu;

pub use crate::inventory::menu::MenuBehavior;

pub use crate::inventory::menu::MenuBuilder;

pub use crate::inventory::menu::MenuKind;

pub use crate::inventory::menu::PlayerInventorySections;

pub use crate::inventory::menu::Rect;

pub use crate::inventory::menu::Region;

pub use crate::inventory::menu::Section;

pub use crate::inventory::menu::SectionKind;

pub use crate::inventory::menu::SectionSource;

pub use crate::inventory::menu::SlotFactory;

pub use crate::inventory::slots::NormalSlot;

pub use crate::inventory::slots::RestrictedSlot;

pub use crate::inventory::slots::ResultHandler;

pub use crate::inventory::slots::Slot;

pub use crate::player::Player;

Structs

ItemStack

A stack of items with a count and component modifications.

Enums

EquipmentSlot

Equipment slots for entities.

EquipmentSlotType

Equipment slot types for categorization.

Traits

IntoShared

Wraps this value in an Arc<SyncMutex<>>

Type Aliases

Shared

A value shared across threads behind the crate’s standard lock.