steel_core::block_entity

Trait BlockEntity

pub trait BlockEntity: Send + Sync {

    // Required methods
    fn as_any(&self) -> &dyn Any;
    fn as_any_mut(&mut self) -> &mut dyn Any;
    fn get_type(&self) -> BlockEntityTypeRef;
    fn get_block_pos(&self) -> BlockPos;
    fn get_block_state(&self) -> BlockStateId;
    fn set_block_state(&mut self, state: BlockStateId);
    fn is_removed(&self) -> bool;
    fn set_removed(&mut self);
    fn clear_removed(&mut self);
    fn get_level(&self) -> Option<Arc<World>>;
    fn load_additional(&mut self, nbt: &BorrowedNbtCompound<'_>);
    fn save_additional(&self, nbt: &mut NbtCompound);

    // Provided methods
    fn set_changed(&self) { ... }
    fn pre_remove_side_effects(&mut self, pos: BlockPos, state: BlockStateId) { ... }
    fn get_update_tag(&self) -> Option<NbtCompound> { ... }
    fn is_ticking(&self) -> bool { ... }
    fn tick(&mut self, world: &Arc<World>) -> Option<BlockEntityTickAction> { ... }
    fn as_container(&self) -> Option<&(dyn Container + 'static)> { ... }
    fn as_container_mut(&mut self) -> Option<&mut (dyn Container + 'static)> { ... }

}

Trait for all block entities.

Block entities are attached to specific blocks in the world and provide additional data storage beyond what block states can hold.

Required Methods

fn as_any(&self) -> &dyn Any

Returns a reference to the block entity as Any for downcasting.

fn as_any_mut(&mut self) -> &mut dyn Any

Returns a mutable reference to the block entity as Any for downcasting.

fn get_type(&self) -> BlockEntityTypeRef

Returns the type of this block entity.

fn get_block_pos(&self) -> BlockPos

Returns the position of this block entity in the world.

fn get_block_state(&self) -> BlockStateId

Returns the current block state associated with this entity.

fn set_block_state(&mut self, state: BlockStateId)

Updates the cached block state.

Called when the block state changes but the block entity is kept.

fn is_removed(&self) -> bool

Returns whether this block entity has been marked for removal.

fn set_removed(&mut self)

Marks this block entity as removed.

Removed block entities will be cleaned up and should not be ticked.

fn clear_removed(&mut self)

Clears the removed flag.

Used when re-adding a block entity that was previously removed.

fn get_level(&self) -> Option<Arc<World>>

Gets the world reference if still valid.

Block entities receive a Weak<World> at construction time.

fn load_additional(&mut self, nbt: &BorrowedNbtCompound<'_>)

Loads additional data from NBT.

Called when loading the block entity from disk or receiving initial chunk data from the server.

fn save_additional(&self, nbt: &mut NbtCompound)

Saves additional data to NBT.

Called when saving the block entity to disk.

Provided Methods

fn set_changed(&self)

Called when the block entity’s data changes.

Marks the containing chunk as dirty so changes are persisted to disk.

fn pre_remove_side_effects(&mut self, pos: BlockPos, state: BlockStateId)

Called before the block entity is removed to handle side effects.

For example, containers should drop their contents here.

Arguments

pos - The position of the block entity

state - The block state being removed

fn get_update_tag(&self) -> Option<NbtCompound>

Returns the NBT data to send to clients for initial sync.

This is included in the chunk data packet when the chunk is first sent. Return None if no client sync is needed.

fn is_ticking(&self) -> bool

Returns whether this block entity should be ticked every game tick.

Block entities that return true will have their tick() method called each game tick.

fn tick(&mut self, world: &Arc<World>) -> Option<BlockEntityTickAction>

Called every game tick for ticking block entities.

Only called if is_ticking() returns true.

fn as_container(&self) -> Option<&(dyn Container + 'static)>

Returns this block entity as a container, if it implements Container.

Override this in block entities that are also containers (e.g., chests, furnaces) to enable integration with the inventory locking system.

fn as_container_mut(&mut self) -> Option<&mut (dyn Container + 'static)>

Returns this block entity as a mutable container, if it implements Container.

Override this in block entities that are also containers (e.g., chests, furnaces) to enable integration with the inventory locking system.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".