steel_core

Module chunk

This module contains all the chunk related structures and logic.

Modules

chunk_access

This module contains the ChunkAccess enum, which is used to access chunks in different states.

chunk_generation_task

This module contains the ChunkGenerationTask struct, which is used to generate chunks.

chunk_generator

This module contains the ChunkGenerator trait, which is used to generate chunks.

chunk_holder

This module contains the ChunkHolder struct, which is used to hold a chunk in a watch channel.

chunk_packet_data

This module contains the ChunkPacketData struct, which is used to create the data for a chunk packet.

level_chunk

This module contains the LevelChunk struct, which is a chunk that is ready to be sent to the client.

paletted_container

A paletted container is a container that can be either homogeneous or heterogeneous.

proto_chunk

A proto chunk is a chunk that is still being generated.

section

This module contains the Sections and ChunkSection structs.