steel_core::chunk::chunk_holder

Struct ChunkHolder

pub struct ChunkHolder {
    chunk_access: Receiver<Option<ChunkStageHolder>>,
    sender: Sender<Option<ChunkStageHolder>>,
}

Holds a chunk in a watch channel, allowing for multiple threads to access it.

Fields
chunk_access: Receiver<Option<ChunkStageHolder>>sender: Sender<Option<ChunkStageHolder>>
Implementations

impl ChunkHolder

pub fn new(proto_chunk: ProtoChunk) -> Self

Creates a new chunk holder.

pub fn with_chunk_mut<F, R>(&self, status: ChunkStatus, f: F) -> Option<R>
where F: FnOnce(&mut ChunkAccess) -> R,

Gets mutable access to the chunk if the chunk has reached the given status.

pub fn with_chunk<F, R>(&self, status: ChunkStatus, f: F) -> Option<R>
where F: FnOnce(&ChunkAccess) -> R,

Gets access to the chunk if the chunk has reached the given status.

pub fn with_full_chunk<F, R>(&self, f: F) -> Option<R>
where F: FnOnce(&LevelChunk) -> R,

Gets access to the chunk if the chunk is full.

Will return None if the chunk is not full or is cancelled.

pub async fn await_chunk_and_then<F, R>( &self, status: ChunkStatus, f: F, ) -> Option<R>
where F: FnOnce(&ChunkAccess) -> R,

Waits until the chunk has reached the given status, and then calls the given function.

pub async fn await_full_and_then<F, R>(&self, f: F) -> Option<R>
where F: FnOnce(&ChunkAccess) -> R,

Waits until this chunk has reached the Full status.

Will return None if the chunk generation is cancelled.

pub async fn await_full(&self) -> Option<()>

Waits until this chunk has reached the Full status.

pub fn upgrade_to_full(&self)

Upgrades the chunk to a full chunk.

Panics

The function expects that the chunk is completed and at ProtoChunk stage