steel_core::command::arguments

Trait CommandArgument

pub trait CommandArgument: Send + Sync {
    type Output;

    // Required methods
    fn parse<'a>(
        &self,
        arg: &'a [&'a str],
        context: &mut CommandContext,
    ) -> Option<(&'a [&'a str], Self::Output)>;
    fn usage(&self) -> (ArgumentType, Option<SuggestionType>);
}

A trait that defines a command argument parser.

Required Associated Types

type Output

The type of the parsed output.

Required Methods

fn parse<'a>( &self, arg: &'a [&'a str], context: &mut CommandContext, ) -> Option<(&'a [&'a str], Self::Output)>

Parses from the given arguments the expected type and returns the remaining unconsumed arguments and the parsed output.

fn usage(&self) -> (ArgumentType, Option<SuggestionType>)

Returns the parser ID associated with this argument.