steel_core::command::commands

Trait CommandExecutor

pub trait CommandExecutor<S> {
    // Required method
    fn execute(
        &self,
        parsed: S,
        context: &mut CommandContext,
    ) -> Result<(), CommandError>;
}

A trait that defines the behavior of a type safe command executor.

Required Methods

fn execute( &self, parsed: S, context: &mut CommandContext, ) -> Result<(), CommandError>

Executes the command with the given type safe arguments.