steel_core::command::commands

Trait CommandHandlerDyn

pub trait CommandHandlerDyn {
    // Required methods
    fn names(&self) -> &'static [&'static str];
    fn description(&self) -> &'static str;
    fn permission(&self) -> &'static str;
    fn execute(
        &self,
        command_args: &[&str],
        context: &mut CommandContext,
        server: &Arc<Server>,
    ) -> Result<(), CommandError>;
    fn usage(&self, buffer: &mut Vec<CommandNode>, root_children: &mut Vec<i32>);
}

Defines a command handler that can be dynamically dispatched.

Required Methods

fn names(&self) -> &'static [&'static str]

Returns the names of the command.

fn description(&self) -> &'static str

Returns the description of the command.

fn permission(&self) -> &'static str

Returns the permission of the command.

fn execute( &self, command_args: &[&str], context: &mut CommandContext, server: &Arc<Server>, ) -> Result<(), CommandError>

Handles the execution of a command sent by a player.

fn usage(&self, buffer: &mut Vec<CommandNode>, root_children: &mut Vec<i32>)

Generates the usage information for the command.