steel_core::entity

Trait LivingEntity

pub trait LivingEntity: Entity {

    // Required methods
    fn get_health(&self) -> f32;
    fn set_health(&self, health: f32);
    fn get_max_health(&self) -> f32;
    fn living_base(&self) -> &SyncMutex<LivingEntityBase>;
    fn get_absorption_amount(&self) -> f32;
    fn set_absorption_amount(&self, amount: f32);
    fn get_armor_value(&self) -> i32;
    fn set_sprinting(&self, sprinting: bool);
    fn get_speed(&self) -> f32;
    fn set_speed(&self, speed: f32);

    // Provided methods
    fn heal(&self, amount: f32) { ... }
    fn is_dead_or_dying(&self) -> bool { ... }
    fn is_alive(&self) -> bool { ... }
    fn is_affected_by_potions(&self) -> bool { ... }
    fn attackable(&self) -> bool { ... }
    fn is_using_item(&self) -> bool { ... }
    fn is_blocking(&self) -> bool { ... }
    fn is_fall_flying(&self) -> bool { ... }
    fn is_sleeping(&self) -> bool { ... }
    fn stop_sleeping(&self) { ... }
    fn is_sprinting(&self) -> bool { ... }
    fn set_delta_movement(&self, velocity: DVec3) { ... }
    fn get_delta_movement(&self) -> DVec3 { ... }

}

A trait for living entities that can take damage, heal, and die.

This trait provides the core functionality for entities that have health, can be damaged, and can die. It’s based on Minecraft’s LivingEntity class.

Note: All methods take &self (not &mut self) because living entities are shared via Arc and use interior mutability (atomics, SyncMutex, etc.).

Required Methods

fn get_health(&self) -> f32

Gets the current health of the entity.

fn set_health(&self, health: f32)

Sets the health of the entity, clamped between 0 and max health.

fn get_max_health(&self) -> f32

Gets the maximum health of the entity.

fn living_base(&self) -> &SyncMutex<LivingEntityBase>

Returns a reference to the shared LivingEntityBase that holds dead, invulnerable_time, and last_hurt.

fn get_absorption_amount(&self) -> f32

Gets the absorption amount (extra health from effects like absorption).

fn set_absorption_amount(&self, amount: f32)

Sets the absorption amount.

fn get_armor_value(&self) -> i32

Gets the entity’s armor value.

fn set_sprinting(&self, sprinting: bool)

Sets whether the entity is sprinting.

fn get_speed(&self) -> f32

Gets the entity’s speed attribute value.

fn set_speed(&self, speed: f32)

Sets the entity’s speed.

Provided Methods

fn heal(&self, amount: f32)

Heals the entity by the specified amount.

fn is_dead_or_dying(&self) -> bool

Returns true if the entity is dead or dying (health <= 0).

fn is_alive(&self) -> bool

Returns true if the entity is alive (health > 0).

fn is_affected_by_potions(&self) -> bool

Checks if the entity can be affected by potions.

fn attackable(&self) -> bool

Checks if the entity is attackable.

fn is_using_item(&self) -> bool

Checks if the entity is currently using an item.

fn is_blocking(&self) -> bool

Checks if the entity is blocking with a shield or similar item.

fn is_fall_flying(&self) -> bool

Checks if the entity is fall flying (using elytra).

fn is_sleeping(&self) -> bool

Checks if the entity is sleeping.

fn stop_sleeping(&self)

Stops the entity from sleeping.

fn is_sprinting(&self) -> bool

Checks if the entity is sprinting.

fn set_delta_movement(&self, velocity: DVec3)

Set the delta_movement of a LivingEntity

fn get_delta_movement(&self) -> DVec3

Get the delta_movement of a LivingEntity