steel_core::player::player_inventory

Struct PlayerInventory

pub struct PlayerInventory {
    items: [ItemStack; 36],
    equipment: EntityEquipment,
    player: Weak<Player>,
    selected: u8,
    times_changed: u32,
}

Player inventory container managing the main inventory and equipment.

Contains 36 main inventory slots (0-8 hotbar, 9-35 main) plus equipment slots (armor, offhand, etc.) accessed through the Container trait.

Fields
items: [ItemStack; 36]

The 36 main inventory slots (0-8 hotbar, 9-35 main).

equipment: EntityEquipment

Entity equipment (armor, hands).

player: Weak<Player>

Weak reference to the player.

selected: u8

Currently selected hotbar slot (0-8).

times_changed: u32

Counter incremented on every change.

Implementations

impl PlayerInventory

pub const INVENTORY_SIZE: usize = 36

Number of main inventory slots.

pub const SELECTION_SIZE: usize = 9

Number of hotbar slots.

pub const SLOT_OFFHAND: usize = 40

Slot index for offhand.

pub fn new(player: Weak<Player>) -> Self

Creates a new player inventory with empty slots.

pub const fn equipment(&self) -> &EntityEquipment

Returns a reference to the entity equipment.

pub const fn equipment_mut(&mut self) -> &mut EntityEquipment

Returns a mutable reference to the entity equipment.

pub const fn is_hotbar_slot(slot: usize) -> bool

Returns true if the given slot index is a hotbar slot (0-8).

pub const fn get_selected_slot(&self) -> u8

Returns the currently selected hotbar slot (0-8).

pub fn set_selected_slot(&mut self, slot: u8)

Sets the selected hotbar slot.

Panics

Panics if the slot is not a valid hotbar slot (must be 0-8).

pub fn with_selected_item<R>(&self, f: impl FnOnce(&ItemStack) -> R) -> R

Executes a function with a reference to the currently selected item.

pub const fn get_selected_item(&self) -> &ItemStack

Returns a clone of the currently selected item (main hand).

pub const fn get_selected_item_mut(&mut self) -> &mut ItemStack

Returns a clone of the currently selected item (main hand).

pub fn set_selected_item(&mut self, item: ItemStack)

Sets the currently selected item (main hand).

pub const fn get_offhand_item(&self) -> &ItemStack

Returns a clone of the offhand item.

pub const fn get_offhand_item_mut(&mut self) -> &mut ItemStack

Returns a clone of the offhand item.

pub fn set_offhand_item(&mut self, item: ItemStack)

Sets the offhand item.

pub fn with_selected_item_mut<R>( &mut self, f: impl FnOnce(&mut ItemStack) -> R, ) -> R

Executes a function with a mutable reference to the currently selected item.

pub const fn get_times_changed(&self) -> u32

Returns the number of times this inventory has been modified.

pub const fn get_items(&self) -> &[ItemStack; 36]

Returns the non-equipment items (main 36 slots).

pub fn get_free_slot(&self) -> i32

Finds the first empty slot in the inventory, or -1 if full.

pub fn find_slot_matching_item(&self, stack: &ItemStack) -> i32

Finds a slot containing an item matching the given stack (same item type). Returns -1 if not found.

pub fn pick_slot(&mut self, slot: i32)

Swaps items between selected hotbar slot and the given slot. Used for pick block when item is in main inventory but not hotbar.

pub fn add_and_pick_item(&mut self, stack: ItemStack) -> bool

Adds an item to the hotbar (for creative pick block) and selects it. Returns true if successful.

pub const fn get_item_in_hand(&self, hand: InteractionHand) -> &ItemStack

Gets the item in the specified hand.

pub const fn get_item_in_hand_mut( &mut self, hand: InteractionHand, ) -> &mut ItemStack

Gets the item in the specified hand.

pub fn set_item_in_hand(&mut self, hand: InteractionHand, item: ItemStack)

Sets the item in the specified hand.